+++ packages/amplify-category-geo/amplify-plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "geo",
   "type": "category",
-  "commands": ["add", "push", "remove", "update", "help", "console", "populate"],
+  "commands": ["add", "push", "remove", "update", "help", "console", "import"],
   "commandAliases": {
     "configure": "update"
   },
